feat: add Modules tasks
