+++ src/fs/copy.js
@@ -1,5 +1,5 @@
 import {cp} from "fs";
-import {getDirname} from "../utils/getDirname.js";
+import {getDirname} from "./utils/getDirname.js";
 import ApiError from "./apiError/apiError.js";
 
 export const copy = async () => {
