+++ src/fs/list.js
@@ -1,5 +1,5 @@
 import {readdir} from 'fs';
-import {getDirname} from "../utils/getDirname.js";
+import {getDirname} from "./utils/getDirname.js";
 import ApiError from "./apiError/apiError.js";
 
 export const list = async () => {
