+++ src/fs/read.js
@@ -1,6 +1,6 @@
 import {readFile} from 'fs';
 import ApiError from "./apiError/apiError.js";
-import {getDirname} from "../utils/getDirname.js";
+import {getDirname} from "./utils/getDirname.js";
 
 export const read = async () => {
     await readFile(`${getDirname(import.meta.url)}/files/fileToRead.txt`,"utf8",
