+++ src/fs/rename.js
@@ -1,5 +1,5 @@
 import {access, rename as renameFile} from 'fs'
-import {getDirname} from "../utils/getDirname.js";
+import {getDirname} from "./utils/getDirname.js";
 import ApiError from "./apiError/apiError.js";
 
 export const rename = async () => {
